refactor: 取引検索ロジックをserver actionからloaderに移動
読み取り専用の検索ロジックはactionではなくloaderに配置すべきため、
loadTransactionsByNosローダーを作成しAPI route経由で呼び出す形に変更。
actionファイルには副作用のある削除処理のみを残す。
